Change timeout argument to int

+++ mcstatus/server.py
@@ -33,10 +33,10 @@ class MCServer(ABC):
 
     :param str host: The host/ip of the minecraft server.
     :param int port: The port that the server is on.
-    :param float timeout: The timeout in seconds before failing to connect.
+    :param int timeout: The timeout in seconds before failing to connect.
     """
 
-    def __init__(self, host: str, port: int, timeout: float = 3):
+    def __init__(self, host: str, port: int, timeout: int = 3):
         self.address = Address(host, port)
         self.timeout = timeout
 
@@ -51,7 +51,7 @@ class MCServer(ABC):
         return self.address.port
 
     @classmethod
-    def lookup(cls, address: str, timeout: float = 3) -> Self:
+    def lookup(cls, address: str, timeout: int = 3) -> Self:
         """Mimics minecraft's server address field.
 
         :param str address: The address of the Minecraft server, like `example.com:19132`
@@ -64,12 +64,12 @@ class MCServer(ABC):
 class JavaServer(MCServer):
     """Base class for a Minecraft Java Edition server."""
 
-    def __init__(self, host: str, port: int = 25565, timeout: float = 3):
+    def __init__(self, host: str, port: int = 25565, timeout: int = 3):
         """Override init to add a default port for java servers of 25565."""
         super().__init__(host, port, timeout=timeout)
 
     @classmethod
-    def lookup(cls, address: str, timeout: float = 3) -> Self:
+    def lookup(cls, address: str, timeout: int = 3) -> Self:
         """Mimics minecraft's server address field.
 
         With Java servers, on top of just parsing the address, we also check the
@@ -84,7 +84,7 @@ class JavaServer(MCServer):
         return cls(addr.host, addr.port, timeout=timeout)
 
     @classmethod
-    async def async_lookup(cls, address: str, timeout: float = 3) -> Self:
+    async def async_lookup(cls, address: str, timeout: int = 3) -> Self:
         """Asynchronous alternative to lookup
 
         For more details, check the docstring of the synchronous lookup function.
@@ -210,7 +210,7 @@ class JavaServer(MCServer):
 class BedrockServer(MCServer):
     """Base class for a Minecraft Bedrock Edition server."""
 
-    def __init__(self, host: str, port: int = 19132, timeout: float = 3):
+    def __init__(self, host: str, port: int = 19132, timeout: int = 3):
         """Override init to add a default port for bedrock servers of 19132."""
         super().__init__(host, port, timeout=timeout)
 
@@ -242,7 +242,7 @@ class MinecraftServer(JavaServer):
     This class is kept purely for backwards compatibility reasons and will be removed eventually.
     """
 
-    def __init__(self, host: str, port: int = 25565, timeout: float = 3):
+    def __init__(self, host: str, port: int = 25565, timeout: int = 3):
         super().__init__(host, port=port, timeout=timeout)
 
 
@@ -253,5 +253,5 @@ class MinecraftBedrockServer(BedrockServer):
     This class is kept purely for backwards compatibility reasons and will be removed eventually.
     """
 
-    def __init__(self, host: str, port: int = 19132, timeout: float = 3):
+    def __init__(self, host: str, port: int = 19132, timeout: int = 3):
         super().__init__(host, port=port, timeout=timeout)
